feat: fix the screen flicker
