+++ packages/shared/src/components/modals/NewSquadModal.tsx
@@ -13,6 +13,7 @@ import {
   SquadStateProps,
 } from '../squads/utils';
 import { SquadComment } from '../squads/Comment';
+import { SquadsConfirmation } from './SquadsConfirmationModal';
 
 export const modalStateOrder = [
   ModalState.Details,
@@ -24,14 +25,18 @@ export const modalStateOrder = [
 export type NewSquadModalProps = {
   onRequestClose: () => void;
   onPreviousState: () => void;
+  isOpen: boolean;
 };
 
 function NewSquadModal({
   onPreviousState,
   onRequestClose,
+  isOpen,
 }: NewSquadModalProps): ReactElement {
   const [modalState, setModalState] = useState(ModalState.Details);
   const [squad, setSquad] = useState<Squad>();
+  const [showConfirmation, setShowConfirmation] = useState(false);
+
   const changeState = (prev: boolean) => {
     const index = modalStateOrder.findIndex((state) => state === modalState);
     const changeTo = modalStateOrder[index + (prev ? -1 : 1)];
@@ -60,32 +65,40 @@ function NewSquadModal({
     onNext,
   };
   return (
-    <Modal
-      isOpen
-      kind={Modal.Kind.FixedCenter}
-      size={Modal.Size.Small}
-      onRequestClose={onRequestClose}
-    >
-      {modalState === ModalState.Ready && <Modal.Header title={modalState} />}
-      {modalState !== ModalState.Ready && (
-        <Modal.Header>
-          <Button
-            icon={<ArrowIcon className="-rotate-90" />}
-            className="flex justify-center items-center mr-2 -ml-2 btn btn-tertiary iconOnly"
-            onClick={() => changeState(true)}
-          />
-          <h3 className="font-bold typo-callout">{modalState}</h3>
-        </Modal.Header>
+    <>
+      <Modal
+        isOpen={isOpen}
+        kind={Modal.Kind.FixedCenter}
+        size={Modal.Size.Small}
+        onRequestClose={() => setShowConfirmation(true)}
+      >
+        {modalState === ModalState.Ready && <Modal.Header title={modalState} />}
+        {modalState !== ModalState.Ready && (
+          <Modal.Header>
+            <Button
+              icon={<ArrowIcon className="-rotate-90" />}
+              className="flex justify-center items-center mr-2 -ml-2 btn btn-tertiary iconOnly"
+              onClick={() => changeState(true)}
+            />
+            <h3 className="font-bold typo-callout">{modalState}</h3>
+          </Modal.Header>
+        )}
+        <div
+          className="absolute left-0 h-1 top-[3.3rem] bg-theme-color-cabbage transition-[width]"
+          style={{ width: `${stepperWidth()}%` }}
+        />
+        <SquadDetails {...stateProps} />
+        <SquadSelectArticle {...stateProps} />
+        <SquadComment {...stateProps} />
+        <SquadReady {...stateProps} squad={squad} />
+      </Modal>
+      {showConfirmation && (
+        <SquadsConfirmation
+          onContinue={onRequestClose}
+          onRequestClose={() => setShowConfirmation(false)}
+        />
       )}
-      <div
-        className="absolute left-0 h-1 top-[3.3rem] bg-theme-color-cabbage transition-[width]"
-        style={{ width: `${stepperWidth()}%` }}
-      />
-      <SquadDetails {...stateProps} />
-      <SquadSelectArticle {...stateProps} />
-      <SquadComment {...stateProps} />
-      <SquadReady {...stateProps} squad={squad} />
-    </Modal>
+    </>
   );
 }
